Moved media to components to keep relative URL valid in build
