+++ components/App.js
@@ -34,8 +34,8 @@ var portfolioItems = [
       'The end of your Lyft ride just got a smooth little update to help you complete payment and rate your driver with a few quick taps. I optimized this experience for the most common use case – you had a fine ride and you’re already calling the next one.',
       'The 3D maps drop off view is a completely new one, and was used specifically to give context about the location of your last ride, and to signal that the ride has ended.',
     ],
-    imgSrc: '../media/rating_payments.png',
-    videoSrcMov: '../media/videos/rating_payment.mov'
+    imgSrc: './media/rating_payments.png',
+    videoSrcMov: './media/videos/rating_payment.mov'
   },
   {
     id: 2,
@@ -44,8 +44,8 @@ var portfolioItems = [
       'We did a complete refresh of Lyft’s  ride experience.  Rooted in rigorous user testing, this update was targted to solve myriad discoverability, accessibility  and usability concerns. I owned the latter half of the experience, including everything after you hit “Request”.',
       'Highlights:<br/>– Contextual ride panel that surfaces the information about your ride when you most need it.<br/>– Ride actions are ergonomically placed within thumbs reach for easy access.',
     ],
-    imgSrc: '../media/rideexperience.png',
-    videoSrcMov: '../media/videos/ride_experience.mov'
+    imgSrc: './media/rideexperience.png',
+    videoSrcMov: './media/videos/ride_experience.mov'
   },
   {
     id: 3,
@@ -53,8 +53,8 @@ var portfolioItems = [
     summary: [
       'Created a subtle, fine-tuned interaction for displaying updated ETAs in the location pin. At once, this animation must effectively communicate the most accurate information without calling too much attentiont to the changes in availability.',
     ],
-    imgSrc: '../media/dynamicpin.png',
-    videoSrcMov: '../media/videos/dynamic_pin_drop.mov'
+    imgSrc: './media/dynamicpin.png',
+    videoSrcMov: './media/videos/dynamic_pin_drop.mov'
   },
   {
     id: 4,
@@ -62,8 +62,8 @@ var portfolioItems = [
     summary: [
       'I created a modular email for Lyft email receipts, to help riders quickly identify the ride, understand pricing, and take action if needed. Components include: ride summary, itemized breakdown of price, and promotional cards. ',
     ],
-    imgSrc: '../media/ride_receipts.png',
-    videoSrcMov: '../media/videos/ride_receipts.mov'
+    imgSrc: './media/ride_receipts.png',
+    videoSrcMov: './media/videos/ride_receipts.mov'
   },
 ]
