+++ components/Avatar.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { StyleSheet, css } from 'aphrodite';
-import vickitan from '../media/vickitan.png';
+import vickitan from './media/vickitan.png';
 
 const styles = StyleSheet.create({
   avatarContainer: {
